WPILibJ Integration Tests
This is the first commit to add the WPILibJ integration test framework.
This creates a project called wpilibJavaIntegrationTests, which is will
create testing framework when running the maven package target. This
framework will need to have the TestBench.java file updated to match
the testing harness set up at WPI. Tests must currently be run manually.

Change-Id: I09fe0278580751e813af591c15b9079360089987