Implements locking in C++ DriverStation, and adds double buffering to DS Thread
The locking currently in the DriverStation is locking on "this", which
is not a good idea. This changes to use a specific lock for joystick
data, and changes that to be a ReadWrite lock. Once the data is cached,
all threads can safely read from the cache at any time. Then when there
is new data, the DS thread grabs an exclusive write lock, caches the
data, and then lets the readers go again.

One thing to note is that this does not fix the inefficiency issues with
the GetData method (it holds the lock for 5 ms out of every 20ms),
however to fix that would require JNI changes in a later patch. Then it
would match the C++ locking changes added.

Change-Id: Id5dfb29b8ac688b9d7a9d4a775cd75bc44233393

Adds caching for DS Joystick Data in Java

Currently, about 5ms of every 20ms loop the DS thread would hold
the mutex, while grabbing data. During this time, and call to grab
joystick data would be blocked. This change grabs the joystick data
to a cache, and then grabs the mutex and moves the data references
around. This is much more efficient then the old code, and gives
teams more of their teleop loop time back for use.

Another major change this does is use preallocated arrays when entering
the JNI. Previously every JNI DS call would allocate an entire new array.
With a GC'd language where those arrays go on the heap, thats a problem,
and creates tons of garbage. That garbage is no longer created anymore,
as all arrays and byte buffers sent to JNI in the DS are preallocated.

Change-Id: I4b40d4409a6199b6bcdfaa68bb5f2d68159ef728

Implements a ReadWrite Lock for C++ DS Joysticks

There is currently no locking on DS joystick data, which means that it
is easily possible for teams to get bad data. In addition, makes all get
data methods used the cached data (all data was being cached, however
some methods were not using it.)

In addition, GetJoystickName was always returning joystick 0 data, which
this fixes.

Change-Id: I2055b66f8f6ffc3e4970f7a6ee5234eed91dca20

Adds proper caching for DS Joystick Data

Before, when the locking was added, there would be 5ms out of every 20
where the DS would hold the lock, and no joystick data could be
accessed. This change grabs new joystick data in separate memory, and
then grabs the lock to update the actual data. This results in much less
contention for locks, and much less potential for user code to be
waiting for data.

The Joystick data structures had to be changed from static arrays to
dynamic arrays in order to save copies, however I made sure to free the
memory when the driver station dies.

Change-Id: I628502c1179da0d9d0182264dd2c9cb395b9b058

Changes ReadWriteLocks to standard Mutex's