Adds safe serial port write methods
The old method had a fairly large race condition, and the way the docs
were written could cause users to get confused. Up for discussion on
adding the old API back and deprecating it, however I think this is
probably a change were we want a full break. The break will also affect
only teams sending a std::string directly to the API, and will most
likely fix teams trying to send a char* to the API.