Adds some functionality to use the new FRC compiler define
I want to add a few more. The current ones are removing all definitions
from the MockData headers when in simulation mode. And a constexpr
IsReal and IsSimulation in RobotBase. That's not as helpful now because
of no if constexpr, but still better then what we had, and useful for
the future.