Improve style conformity via IDEA's "Reformat Code" tool
Imported intellij-java-google-style.xml from the style guide repo.

As an aside, I find it funny that the passage "If you are looking for some easy ramp-up tasks, finding areas that don't follow the style guide and fixing them is very welcome." is present in CONTRIBUTING.md, because this took 5 minutes and no effort.