Added delta tolerance term to PIDController's SetTolerance()
PIDController returns true from OnTarget() when the output is passing through
the reference, not only when the output has settled. To compensate for this,
teams have been using the tolerance buffer to lag the OnTarget() calculation
until the oscillations have settled. This is obviously not the intended use for
the tolerance buffer.

This patch adds a defaulted argument to SetTolerance() for the change in error
between samples. This delta tolerance term ensures the controller is not
considered to be at the reference when oscillating around it at large
frequencies.