Silence DriverStation Joystick Unplugged Warnings. Closes #962 Only suppressed the reportJoystickUnpluggedWarning because the Error counterpart is used to alert for the index error. Something that could be a legitimate issue.
I put it into the time check condition, but put it first so it can short circuit faster if enabled.

silenceJoystickUnplugged(boolean silenced) allows a toggling as needed (for example, teams having a debugging controller that prints certain messages)

silenceJoystickUnplugged() enables silencing.

getJoystickWarngins() - Get whether the warnings are suppressed.
I couldn't think of a good name for this, as the others I had in mind were either cumbersome or not grammatically correct.